requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
pymongo>=4.0.0
redis>=5.0.0
rethinkdb>=2.4.0
PyYAML>=6.0
python-dotenv>=1.0.0
httpx>=0.24.0
numpy>=1.24.0
scikit-learn>=1.3.0